+++ docs/CONTRIBUTION.md
@@ -43,5 +43,4 @@ We welcome contributions to our project! Here's how you can help:
 3. Increase the version numbers in any examples files and the README.md to the new version that this Pull Request would represent
 4. You may merge the Pull Request in once you have the sign-off of two other developers, or if you do not have permission to do that, you may request the second reviewer to merge it for you
 
-Thank you for your contributions! 
-
+Thank you for your contributions!
