+++ src/app.py
@@ -1,5 +1,5 @@
 import streamlit as st
-from backend.modules.css import load_css
+from src.backend.config.css import load_css
 
 # Pages
 home = st.Page(
